+++ app/components/layout/SidebarAdmin.vue
@@ -1,7 +1,7 @@
 <template>
-  <aside class="w-14.5 m-5 h-[calc(100vh-40px)] m-5 bg-white rounded-3xl flex flex-col items-center py-8">
+  <aside class="w-20 m-5 h-[calc(100vh-40px)] bg-white rounded-3xl flex flex-col items-center py-8">
     
-    <img src="~/assets/images/logo.png" alt="Logo" class="w-8 h-8 mb-10 shrink-0" />
+    <img src="~/assets/images/logo.png" alt="Logo" class="w-10 h-10 mb-10 shrink-0" />
     
     <nav class="flex flex-col gap-4 flex-1 w-full px-3">
       <NavIcon icon-name="i-lucide-home" class="w-full" is-active/>