+++ app/components/nav/Button.vue
@@ -8,7 +8,7 @@ const props = withDefaults(defineProps<{
 
 const classes = computed(() => [
   'px-4 py-2 rounded-[16px]',
-  'font-bold Montserrat text-xs text-black',
+  'font-bold text-black',
   'transition-all duration-200',
   props.isActive ? 'bg-[#DFE3F0]' : 'bg-transparent hover:bg-gray-100'
 ])