+++ app/components/nav/Icon.vue
@@ -17,11 +17,11 @@ const classes = computed(() => [
 
 <template>
   <button :class="classes">
-    <UIcon :name="iconName" class="size-5 shrink-0" />
+    <UIcon :name="iconName" class="size-6 shrink-0" />
     
     <span 
       v-if="badgeCount > 0" 
-      class="absolute -top-0.5 -right-0.5 bg-[#C9353F] text-white text-[11px] font-bold min-w-[20px] h-[20px] px-1 flex items-center justify-center rounded-full"
+      class="absolute -top-0.5 -right-0.5 bg-[#C9353F] text-white text-[11px] font-bold min-w-5 h-5 px-1 flex items-center justify-center rounded-full"
     >
       {{ badgeCount }}
     </span>